FaceEmoteControlProcessorへ処理を分離
