+++ Editor/FaceEmoteControlProcessor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ebde9e9e97b1a844fbf744f05afc41f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
